+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Version 0.11alpha
 -----------------
 - support for column grants
 - function ddlx_get_dependants_recursive() rolled into ddlx_get_dependants()
+- support for fdw options on columns
 
 Version 0.10
 ------------